medisebi/backend/requirements.txt
# ── MediSebi Backend Dependencies ───────────────────────────────
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20

# AI/ML
scikit-learn==1.6.1
prophet==1.1.5
pandas==2.2.3
numpy==1.26.4

# External APIs
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
matplotlib==3.10.0
